removed mysql as an input data type, fix #127
