+++ docs/api/smrf.data.rst
@@ -4,34 +4,50 @@ smrf.data package
 Submodules
 ----------
 
-smrf.data.loadData module
--------------------------
+smrf.data.csv module
+--------------------
 
-.. automodule:: smrf.data.loadData
+.. automodule:: smrf.data.csv
    :members:
    :undoc-members:
    :show-inheritance:
 
-smrf.data.loadGrid module
--------------------------
+smrf.data.hrrr\_grib module
+---------------------------
 
-.. automodule:: smrf.data.loadGrid
+.. automodule:: smrf.data.hrrr_grib
    :members:
    :undoc-members:
    :show-inheritance:
 
-smrf.data.loadTopo module
--------------------------
+smrf.data.load\_data module
+---------------------------
 
-.. automodule:: smrf.data.loadTopo
+.. automodule:: smrf.data.load_data
    :members:
    :undoc-members:
    :show-inheritance:
 
-smrf.data.mysql\_data module
-----------------------------
+smrf.data.load\_topo module
+---------------------------
 
-.. automodule:: smrf.data.mysql_data
+.. automodule:: smrf.data.load_topo
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+smrf.data.netcdf module
+-----------------------
+
+.. automodule:: smrf.data.netcdf
+   :members:
+   :undoc-members:
+   :show-inheritance:
+
+smrf.data.wrf module
+--------------------
+
+.. automodule:: smrf.data.wrf
    :members:
    :undoc-members:
    :show-inheritance:
