+++ requirements.txt
@@ -1,7 +1,6 @@
 coloredlogs
 Cython>=0.28.4
 inicheck>=0.9.0,<0.10.0
-mysql-connector-python-rf==2.2.2
 netCDF4>=1.2.9
 numpy>=1.14.0,<1.19.0
 pandas>=0.23.0
