+++ smrf/framework/CoreConfig.ini
@@ -104,75 +104,6 @@ cloud_factor:
 type = CriticalFilename,
 description = Path to CSV containing the station measured cloud factor
 
-################################################################################
-# MySQL database configurations
-################################################################################
-
-[mysql]
-
-stations:
-type = station list,
-description = List of station IDs to use for distributing any of the variables
-
-user:
-description = username for database login.
-
-password:
-type = password,
-description = password used for database login.
-
-host:
-description = IP address to server.
-
-port:
-default = 3606,
-type = int,
-description = Port for MySQL database.
-
-database:
-default = weather_db,
-description = name of the database containing station data
-
-data_table:
-default = tbl_level2,
-description = name of the database table containing station data
-
-station_table:
-default = tbl_stations,
-description = name of the database table containing client and source
-
-metadata:
-default = tbl_metadata,
-description = name of the database table containing station metadata
-
-air_temp:
-default = air_temp,
-description = name of the table column containing station air temperature
-
-vapor_pressure:
-default = vapor_pressure,
-description = name of the table column containing station vapor pressure
-
-precip:
-default = precip_accum,
-description = name of the table column containing station precipitation
-
-wind_speed:
-default = wind_speed,
-description = name of the table column containing station wind speed
-
-wind_direction:
-default = wind_direction,
-description = name of the table column containing station wind direction
-
-cloud_factor:
-default = cloud_factor,
-description = name of the table column containing station cloud factor
-
-solar:
-default = solar_radiation,
-description = name of the table column containing station solar radiation
-
 
 ################################################################################
 # Gridded datasets configurations
