+++ smrf/framework/recipes.ini
@@ -215,9 +215,6 @@ trigger:
 csv:
   apply_defaults = true
 
-mysql:
-  remove_section = true
-
 gridded:
   remove_section = true
 
@@ -227,28 +224,9 @@ thermal:
 air_temp:
   remove_item = zach
 
-[mysql_recipe]
-trigger:
-  has_section = mysql
-
-mysql:
-  apply_defaults = true
-
-csv:
-  remove_section = true
-
-gridded:
-  remove_section = true
-
-thermal:
-  remove_item = [distribution detrend detrend_slope grid_method]
-
 [remove_wind_ninja_recipe]
 trigger:
   has_section = csv
-
-trigger_csv:
-  has_section = mysql
 
 trigger_wind_interp:
   has_value = [wind wind_model interp]
@@ -263,9 +241,6 @@ gridded:
   apply_defaults = true
 
 csv:
-  remove_section = true
-
-mysql:
   remove_section = true
 
 air_temp:
