+++ smrf/tests/data/test_csv.py
@@ -1,103 +1,8 @@
-#!/usr/bin/env python
-# -*- coding: utf-8 -*-
-
-"""
-test_load_data
-----------------------------------
-
-Tests for `data.load_data` module.
-"""
-
-from copy import deepcopy
 
 from inicheck.tools import cast_all_variables
 
 from smrf.framework.model_framework import run_smrf, SMRF
-# from smrf.tests.nwrc_check import NWRCCheck
 from smrf.tests.smrf_test_case import SMRFTestCase
-
-# @unittest.skipUnless(
-#     NWRCCheck.in_network(),
-#     "Skipping because we are not on the NWRC network"
-# )
-# class TestLoadMySQLData(SMRFTestCase):
-
-#     def test_mysql_data_w_stations(self):
-#         """
-#         Use a simple user tester on the weather database to ensure loading is
-#         performed correctly. This will not work outside of NWRC until we
-#         convert so SQLalchemy.
-#         """
-#         # test a successful run specifying stations
-#         config = deepcopy(self.base_config)
-#         options = deepcopy(NWRCCheck.MYSQL_OPTIONS)
-#         config.raw_cfg['mysql'] = options
-
-#         config.raw_cfg['mysql']['stations'] = ['RMESP', 'RME_176']
-#         del config.raw_cfg['csv']
-
-#         config.apply_recipes()
-#         config = cast_all_variables(config, config.mcfg)
-
-#         result = self.can_i_run_smrf(config)
-
-#         self.assertTrue(result)
-
-#     def test_mysql_data_w_client(self):
-#         """
-#         Run SMRF with MYSQL data from client, also can only be run from
-#  inside
-#         NWRC.
-#         """
-#         # test a successful run specifying client
-#         config = deepcopy(self.base_config)
-#         options = deepcopy(NWRCCheck.MYSQL_OPTIONS)
-#         config.raw_cfg['mysql'] = options
-
-#         config.raw_cfg['mysql']['client'] = 'RME_test'
-#         del config.raw_cfg['csv']
-
-#         config.apply_recipes()
-#         config = cast_all_variables(config, config.mcfg)
-
-#         result = can_i_run_smrf(config)
-#         assert result
-
-#     def test_mysql_metadata_error(self):
-#         """ test no metadata found """
-
-#         config = deepcopy(self.base_config)
-#         options = deepcopy(NWRCCheck.MYSQL_OPTIONS)
-#         config.raw_cfg['mysql'] = options
-
-#         config.raw_cfg['mysql']['stations'] = ['NOT_STID', 'NOPE']
-#         del config.raw_cfg['csv']
-
-#         config.apply_recipes()
-#         config = cast_all_variables(config, config.mcfg)
-
-#         with self.assertRaises(Exception):
-#             run_smrf(config)
-
-#     def test_mysql_data_error(self):
-#         """ test no data found """
-
-#         config = deepcopy(self.base_config)
-#         options = deepcopy(NWRCCheck.MYSQL_OPTIONS)
-
-#         config.raw_cfg['mysql'] = options
-#         config.raw_cfg['mysql']['stations'] = ['RMESP', 'RME_176']
-#         del config.raw_cfg['csv']
-
-#         # wrong time
-#         config.raw_cfg['time']['start_date'] = '1900-01-01 00:00'
-#         config.raw_cfg['time']['end_date'] = '1900-02-01 00:00'
-
-#         config.apply_recipes()
-#         config = cast_all_variables(config, config.mcfg)
-
-#         with self.assertRaises(Exception):
-#             run_smrf(config)
 
 
 class TestLoadCSVData(SMRFTestCase):
@@ -106,7 +11,7 @@ class TestLoadCSVData(SMRFTestCase):
         """
         Test the start date not in the data
         """
-        config = deepcopy(self.base_config)
+        config = self.base_config
 
         # Use dates not in the dataset, expecting an error
         config.raw_cfg['time']['start_date'] = '1900-01-01 00:00'
@@ -125,7 +30,7 @@ class TestLoadCSVData(SMRFTestCase):
         """
 
         # test the end date
-        config = deepcopy(self.base_config)
+        config = self.base_config
         config.raw_cfg['csv']['stations'] = ['RMESP', 'RME_176']
 
         # apply the new recipies
