+++ smrf/tests/distribute/wind/test_wind_ninja.py
@@ -5,7 +5,7 @@ import numpy as np
 import pytz
 
 from smrf.data import Topo
-from smrf.data.mysql_data import date_range
+from smrf.utils.utils import date_range
 from smrf.distribute.wind.wind_ninja import WindNinjaModel
 from smrf.tests.smrf_test_case_lakes import SMRFTestCaseLakes
 from smrf.utils import utils
